refactor: postToHook()をコンポーネントにした
