+++ src/main/java/jp/co/esm/miffy/Controller/Asf4MemberController.java
@@ -1,5 +1,6 @@
 package jp.co.esm.miffy.Controller;
 
+import jp.co.esm.miffy.component.HookComponent;
 import jp.co.esm.miffy.entity.Asf4Member;
 import jp.co.esm.miffy.service.HookService;
 import lombok.RequiredArgsConstructor;
@@ -13,12 +14,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class Asf4MemberController {
     private final HookService hookService;
+    private final HookComponent hookComponent;
 
     @GetMapping("asf4members")
     public String index(Model model) {
         List<Asf4Member> asf4MemberList = hookService.selectAll();
-        model.addAttribute("asf4MemberList", asf4MemberList); // debug用
-        hookService.postToHook();
+        model.addAttribute("asf4MemberList", asf4MemberList);
+        hookComponent.postToHook();
         return "asf4members";
     }
 }
